Add access to a pre-trained model
Predictions made from the model inside the juypter notebook exhibit
acceptable levels of echoing and memorization. For the web app to
work correctly, text generated from a .h5 model should be approach
the same quality.

After tuning manually, a handful of generated texts were produced and
evaluated (by a human). As a result, the model uploaded here produced
an acceptable starting point for deployment.

Some observations to be aware of are:
- The model is still memorizing
- The generated text tends to echo itself
- The generated text inside and outside the notebook are drastically different
	on the same starting string